Fix for reply txt problem, admin patches
git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@172 fff143f6-a9c2-4b89-8f93-78cb2c0259b0

+++ mydns/tags/1.2.8/src/mydns/reply.c
@@ -537,7 +537,7 @@ reply_add_srv(TASK *t, RR *r) {
 static inline int
 reply_add_txt(TASK *t, RR *r) {
   char		*dest = NULL;
-  uint8_t	size = 0;
+  size_t	size = 0;
   size_t	len = 0;
   MYDNS_RR	*rr = (MYDNS_RR *)r->rr;
